+++ package.json
@@ -86,7 +86,6 @@
     "eslint-webpack-plugin": "^3.1.1",
     "file-loader": "^6.2.0",
     "glob": "^7.2.0",
-    "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
     "identity-obj-proxy": "^3.0.0",
     "image-webpack-loader": "^8.1.0",
