+++ scripts/webpack.config.js
@@ -8,7 +8,6 @@ import path from "path";
 import LoadablePlugin from "@loadable/webpack-plugin";
 import CssMinimizerPlugin from "css-minimizer-webpack-plugin";
 import ESLintPlugin from "eslint-webpack-plugin";
-import HtmlWebpackPlugin from "html-webpack-plugin";
 import MiniCssExtractPlugin from "mini-css-extract-plugin";
 import svgToMiniDataURI from "mini-svg-data-uri";
 import StyleLintPlugin from "stylelint-webpack-plugin";
@@ -153,7 +152,7 @@ const config = {
         type: "asset/resource",
         generator: {
           filename: "fonts/[name][ext]",
-          emit: false, // fonts are generated by the server config
+          emit: true,
         },
       },
       {
@@ -163,7 +162,6 @@ const config = {
           /\.(jpe?g|png|gif|svg|webp|avif)$/,
           /\.(woff2?|[ot]tf|eot)$/,
           /\.json$/,
-          /\.hbs$/,
         ],
         type: "asset/resource",
         generator: {
@@ -212,7 +210,7 @@ const clientConfig = {
         rules: [
           {
             loader: isDevelopment
-              ? "style-loader" // style-loader caused duplication of critical styles in development
+              ? "style-loader"
               : MiniCssExtractPlugin.loader,
           },
           {
@@ -352,18 +350,6 @@ const clientConfig = {
         isDevelopment ? "[name].css" : "[name].[contenthash].css"
       }`,
     }),
-    new HtmlWebpackPlugin({
-      filename: path.resolve(OUTPUT_DIR, "index.hbs"),
-      showErrors: isDevelopment,
-      template: path.resolve(SRC_DIR, "templates", "index.hbs"),
-      inject: false,
-    }),
-    new HtmlWebpackPlugin({
-      filename: path.resolve(OUTPUT_DIR, "500.hbs"),
-      showErrors: isDevelopment,
-      template: path.resolve(SRC_DIR, "templates", "500.hbs"),
-      inject: false,
-    }),
     ...(isDevelopment
       ? []
       : [
@@ -415,9 +401,13 @@ const serverConfig = {
       {
         test: /\.(sa|sc|c)ss$/,
         rules: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-          },
+          ...(isDevelopment
+            ? []
+            : [
+                {
+                  loader: MiniCssExtractPlugin.loader,
+                },
+              ]),
           {
             include: SRC_DIR,
             loader: "css-loader",
@@ -427,6 +417,7 @@ const serverConfig = {
                 localIdentName: isDevelopment
                   ? "[name]-[local]-[hash:base64:5]"
                   : "[hash:base64:5]",
+                exportOnlyLocals: isDevelopment,
               },
               importLoaders: 1,
             },
@@ -540,7 +531,7 @@ const serverConfig = {
         type: "asset/resource",
         generator: {
           filename: "fonts/[name][ext]",
-          emit: true,
+          emit: false,
         },
       },
     ],
@@ -548,9 +539,13 @@ const serverConfig = {
   externals: ["@loadable/component", nodeExternals()],
   plugins: [
     ...config.plugins,
-    new MiniCssExtractPlugin({
-      filename: "[name].css",
-    }),
+    ...(isDevelopment
+      ? []
+      : [
+          new MiniCssExtractPlugin({
+            filename: "[name].css",
+          }),
+        ]),
   ],
   optimization: {
     minimizer: [new CssMinimizerPlugin()],
