+++ src/client.js
@@ -9,16 +9,11 @@ import configureStore from "./js/store";
 
 // grab the state from a global variable injected into the server-generated HTML
 // eslint-disable-next-line no-underscore-dangle
-const preloadedState = window.__PRELOADED_STATE__;
-
-// allow the passed state to be garbage-collected
-// eslint-disable-next-line no-underscore-dangle
-delete window.__PRELOADED_STATE__;
-
-const store = configureStore(preloadedState);
-const routes = selectRoutes(store.getState());
+const store = configureStore(window.__PRELOADED_STATE__);
 
 loadableReady(() => {
+  const routes = selectRoutes(store.getState());
+  // eslint-disable-next-line no-console
   ReactDOM.hydrate(
     <BrowserRouter>
       <App store={store}>
