+++ src/js/reducerRegistry.js
